designed the right panel for units and buildings purchase
