+++ src/TinyCiv.Client/Code/ClientSingleton.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using System.Threading;
 using TinyCiv.Server.Client;
-using TinyCiv.Shared.Events;
 
 namespace TinyCiv.Client.Code.MVVM
 {
