+++ src/TinyCiv.Client/Code/MVVM/MainViewModel.cs
@@ -1,16 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TinyCiv.Client.Code.Core;
-using TinyCiv.Shared.Game;
 using TinyCiv.Client.Code.MVVM.ViewModel;
-using TinyCiv.Server.Client;
-using TinyCiv.Shared.Events.Client;
 using System.Windows;
 using TinyCiv.Shared.Events.Server;
-using TinyCiv.Client.Code.Units;
 using System.Threading;
 using TinyCiv.Shared.Events.Client.Lobby;
 
@@ -51,7 +42,6 @@ namespace TinyCiv.Client.Code.MVVM
                 CurrentPlayer.Instance.player = response.Created;
                 UpperMenuVM.PlayerColor.Value = CurrentPlayer.Color;
             }
-            MessageBox.Show($"Player: {CurrentPlayer.Id} has joined the game! They are in the {CurrentPlayer.Color} team!");
 
             // If the party is full
             if (CurrentPlayer.Instance.player == null)
