+++ src/TinyCiv.Client/Code/MVVM/ObservableObject.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
+﻿using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace TinyCiv.Client.Code.Core
+namespace TinyCiv.Client.Code.MVVM
 {
     public class ObservableObject : INotifyPropertyChanged
     {
