+++ src/TinyCiv.Client/Code/MVVM/ObservableValue.cs
@@ -1,13 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading.Tasks;
-using TinyCiv.Client.Code.Core;
-
-namespace TinyCiv.Client.Code.MVVM
+﻿namespace TinyCiv.Client.Code.MVVM
 {
     public class ObservableValue<T> : ObservableObject
     {
