+++ src/TinyCiv.Client/Code/MVVM/View/UnitMenuView.xaml
@@ -3,13 +3,12 @@
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:TinyCiv.Client.Code.MVVM.View"
+      xmlns:local="clr-namespace:TinyCiv.Client.Code.MVVM"             
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="200">
     <UserControl.Resources>
-        <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
+        <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter"/>
     </UserControl.Resources>
-
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="30"></RowDefinition>
@@ -27,25 +26,25 @@
             <Button Grid.Column="1" Content="Buildings" Command="{Binding ShowBuildingsCommand}" />
         </Grid>
 
-        <StackPanel Grid.Row="1" Visibility="{Binding IsUnitsListVisible, Converter={StaticResource BooleanToVisibilityConverter}}">
+        <StackPanel Grid.Row="1" Visibility="{Binding IsUnitsListVisible.Value, Converter={StaticResource BooleanToVisibilityConverter}}">
             <ItemsControl ItemsSource="{Binding UnitList}">
                 <ItemsControl.ItemTemplate>
                     <DataTemplate>
                         <Border BorderBrush="LightGray" BorderThickness="1">
-                            <Grid Height="auto" Margin="10">
+                            <Grid Height="auto" Margin="5">
                                 <Grid.ColumnDefinitions>
-                                    <ColumnDefinition Width="15*" />
-                                    <ColumnDefinition Width="85*"/>
+                                    <ColumnDefinition Width="20*" />
+                                    <ColumnDefinition Width="80*"/>
                                 </Grid.ColumnDefinitions>
                                 <Grid Grid.Column="0">
                                     <Grid.RowDefinitions>
                                         <RowDefinition />
-                                        <RowDefinition Height="auto"/>
+                                        <RowDefinition Height="auto" />
                                     </Grid.RowDefinitions>
-                                    <Image Grid.Row="0" Source="{Binding ImagePath}"></Image>
-                                    <TextBlock Grid.Row="1" HorizontalAlignment="Center" FontSize="6" Text="{Binding Name}"/>
+                                    <Image Grid.Row="0" VerticalAlignment="Center" Source="{Binding ImagePath}"></Image>
+                                    <TextBlock Grid.Row="1" HorizontalAlignment="Center" VerticalAlignment="Bottom" FontSize="8" Text="{Binding Name}"/>
                                 </Grid>
-                                <Grid Grid.Column="1" Margin="10,0,0,0">
+                                <Grid Grid.Column="1" Margin="5,0,0,0">
                                     <Grid.RowDefinitions>
                                         <RowDefinition Height="auto" />
                                         <RowDefinition Height="*"/>
@@ -70,8 +69,8 @@
                                             <ColumnDefinition Width="40*" />
                                             <ColumnDefinition Width="40*"/>
                                         </Grid.ColumnDefinitions>
-                                        <TextBlock Grid.Column="0" FontSize="10">Cost:</TextBlock>
-                                        <TextBlock Grid.Column="1" FontSize="10" Text="{Binding ProductionPrice}" />
+                                        <TextBlock Grid.Column="0" FontSize="10" VerticalAlignment="Bottom">Cost:</TextBlock>
+                                        <TextBlock Grid.Column="1" FontSize="10" VerticalAlignment="Bottom" Text="{Binding ProductionPrice}" />
                                         <Button Grid.Column="2" FontSize="10" >Buy</Button>
                                     </Grid>
                                 </Grid>
@@ -82,26 +81,26 @@
             </ItemsControl>
         </StackPanel>
 
-        <StackPanel Grid.Row="1" Visibility="{Binding IsBuildingsListVisible, Converter={StaticResource BooleanToVisibilityConverter}}">
+        <StackPanel Grid.Row="1" Visibility="{Binding IsBuildingsListVisible.Value, Converter={StaticResource BooleanToVisibilityConverter}}">
             <ItemsControl ItemsSource="{Binding BuildingList}">
                 <ItemsControl.ItemTemplate>
                     <DataTemplate>
                         <Border  BorderBrush="LightGray" BorderThickness="1">
-                            <Grid Margin="10">
+                            <Grid Margin="5">
                                 <Grid.ColumnDefinitions>
-                                    <ColumnDefinition Width="15*" />
-                                    <ColumnDefinition Width="85*"/>
+                                    <ColumnDefinition Width="20*" />
+                                    <ColumnDefinition Width="80*"/>
                                 </Grid.ColumnDefinitions>
                                 <Grid Grid.Column="0">
                                     <Grid.RowDefinitions>
                                         <RowDefinition />
-                                        <RowDefinition Height="auto"/>
+                                        <RowDefinition Height="auto" />
                                     </Grid.RowDefinitions>
-                                    <Image Grid.Row="0" Source="{Binding ImagePath}" />
-                                    <TextBlock Grid.Row="1" HorizontalAlignment="Center" FontSize="8" Text="{Binding Name}" />
+                                    <Image Grid.Row="0" VerticalAlignment="Center" Source="{Binding ImagePath}" />
+                                    <TextBlock Grid.Row="1" HorizontalAlignment="Center" VerticalAlignment="Bottom" FontSize="8" Text="{Binding Name}" />
                                 </Grid>
 
-                                <Grid Grid.Column="1" Margin="10 0 0 0">
+                                <Grid Grid.Column="1" Margin="5 0 0 0">
                                     <Grid.RowDefinitions>
                                         <RowDefinition />
                                         <RowDefinition />
@@ -115,8 +114,8 @@
                                             <ColumnDefinition Width="30*"/>
                                             <ColumnDefinition Width="35*"/>
                                         </Grid.ColumnDefinitions>
-                                        <TextBlock Grid.Column="0" FontSize="10">Cost:</TextBlock>
-                                        <TextBlock Grid.Column="1" FontSize="10" Text="{Binding Cost}"/>
+                                        <TextBlock Grid.Column="0" FontSize="10" VerticalAlignment="Bottom">Cost:</TextBlock>
+                                        <TextBlock Grid.Column="1" FontSize="10" VerticalAlignment="Bottom" Text="{Binding Cost}"/>
                                         <Button Grid.Column="2" FontSize="10">Buy</Button>
                                     </Grid>
                                 </Grid>
