+++ src/TinyCiv.Client/Code/MVVM/ViewModel/GameViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using TinyCiv.Client.Code.Core;
 using TinyCiv.Shared;
 using TinyCiv.Shared.Events.Server;
