+++ src/TinyCiv.Client/Code/MVVM/ViewModel/LobbyMenuViewModel.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Windows.Input;
-using TinyCiv.Client.Code.Core;
+﻿using System.Threading;
 using TinyCiv.Shared.Events.Client;
 using TinyCiv.Shared.Events.Server;
 using TinyCiv.Shared.Game;
@@ -14,10 +7,16 @@ namespace TinyCiv.Client.Code.MVVM.ViewModel
 {
     public class LobbyMenuViewModel : ObservableObject
     {
-        public ObservableValue<RelayCommand> StartGame { get; } = new ObservableValue<RelayCommand>(new RelayCommand( o => 
+        //public ObservableValue<RelayCommand> StartGame { get; } = new ObservableValue<RelayCommand>(new RelayCommand( o => 
+        //{
+        //    ClientSingleton.Instance.serverClient.SendAsync(new StartGameClientEvent(MapType.Watery));
+        //}));
+
+        public RelayCommand StartGame { get; } = new RelayCommand(execute =>
         {
             ClientSingleton.Instance.serverClient.SendAsync(new StartGameClientEvent(MapType.Watery));
-        }));
+        });
+
 
         public ObservableValue<bool> IsLobbyReady { get; } = new ObservableValue<bool>(false);
