+++ src/TinyCiv.Client/Code/MVVM/ViewModel/UpperMenuViewModel.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TinyCiv.Client.Code.Core;
-using TinyCiv.Shared.Game;
+﻿using TinyCiv.Shared.Game;
 
 namespace TinyCiv.Client.Code.MVVM.ViewModel
 {
