+++ src/TinyCiv.Client/MainWindow.xaml.cs
@@ -1,18 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Media.Imaging;
-using TinyCiv.Client.Code;
-using TinyCiv.Client.Code.MVVM;
-using TinyCiv.Client.Code.Units;
-using TinyCiv.Server.Client;
-using TinyCiv.Shared;
-using TinyCiv.Shared.Events.Client;
-using TinyCiv.Shared.Events.Server;
-using TinyCiv.Shared.Game;
+﻿using System.Windows;
 
 namespace TinyCiv.Client
 {
